requirements.txt
streamlit==1.28.0
langgraph==0.0.40
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
numpy==1.24.3
plotly==5.15.0
python-docx==0.8.11
lxml==4.9.3
openpyxl==3.1.2
scikit-learn==1.3.0
jieba==0.42.1
PyMuPDF==1.23.8
zhipuai==2.0.1
langchain-core==0.1.0
langchain-openai==0.0.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
nltk==3.8.1
langchain-community==0.0.10
